fix json unmarshal bug and refactor code
